Isolating  OpenGL windowing framework toolkit.
